refactor: break into modules
