--- packages/imt-rs/src/main.rs
@@ -1,3 +1,0 @@
-fn main(){
-    println!("IMT Rust implementation");
-}
